+++ test/11-createCollection.js
@@ -36,4 +36,21 @@ describe('createCollection', function () {
       })
   })
 
+  it('should forced you to rebuild associated table', function (done) {
+    const cls = new Cls(lib.options)
+    cls.createCollection({ 
+      name: 'newtable',
+      fields: [
+        { id: 'id', type: 'string', default: 'myid', required: true },
+        { id: 'title', type: 'string', required: true },
+        { id: 'draft', type: 'boolean', default: true },
+        { id: 'content', type: 'text', required: true }
+      ]
+    }, { withSchema: true })
+      .then(result => {
+        expect(result).to.have.property('success', true)
+        done()
+      })
+  })
+
 })